tests/wpt/tests/dom/events/non-cancelable-when-passive/generic-events-stay-cancelable.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
async_test((t) => {
    const et = new EventTarget();
    et.addEventListener('test', t.step_func_done((e) => {
        assert_true(e.cancelable);
    }), {passive: true});
    et.dispatchEvent(new Event('test', {cancelable: true}));
}, "A generic event with only passive listeners remains cancelable");
</script>
